+++ packages/gov-ui-kit/src/core/components/accordion/accordionItemHeader/index.ts
@@ -1,0 +1,1 @@
+export { AccordionItemHeader, type IAccordionItemHeaderProps } from './accordionItemHeader';